add debug for document
